Extract unshadow logic to utils file
